% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  projectSettingsFile = NULL,
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{projectSettingsFile}{Optional path to a project YAML file to load at startup.}

\item{...}{Additional arguments to pass to `golem_opts`. See `?golem::get_golem_options` for more details.}
}
\value{
No return value, called for side effects. Launches the Shiny app.
}
\description{
`run_app()` starts the Shiny application contained in this package, with an option to load settings from a YAML file using `projectSettingsFile = "pathToYamlFile"`.
}
\details{
This function initializes and runs the Shiny app developed with the golem framework.
It optionally loads application-specific settings from a YAML file, which can be useful for project-specific configurations.
Project YAML files can be exported from the Settings panel inside the app.
}
\examples{
if (interactive()) {
  # Example: Run the application with default settings

  options(shiny.port=httpuv::randomPort(),
          shiny.launch.browser = .rs.invokeShinyWindowExternal,
          shiny.maxRequestSize=9000*1024^2)

  run_app()

  # Example: Run the application with a project YAML exported from Settings
  run_app(projectSettingsFile = "C:/test-project.yml")
}
}
